backend/requirements.txt
amqp==5.0.6
appdirs==1.4.4
asgiref==3.3.4
billiard==3.6.4.0
black==21.5b1
celery==5.0.5
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
click==7.1.2
click-didyoumean==0.0.3
click-plugins==1.1.1
click-repl==0.1.6
colorama==0.4.4
cryptography==3.4.7
defusedxml==0.7.1
Django==3.2.3
djangorestframework==3.12.4
idna==2.10
jwcrypto==0.8
kombu==5.1.0
mypy-extensions==0.4.3
oauthlib==3.1.0
pathspec==0.8.1
prompt-toolkit==3.0.18
psycopg2==2.8.6
pycparser==2.20
PyJWT==2.1.0
python-decouple==3.4
python3-openid==3.2.0
pytz==2021.1
regex==2021.4.4
requests==2.25.1
requests-oauthlib==1.3.0
six==1.16.0
sqlparse==0.4.1
toml==0.10.2
urllib3==1.26.4
vine==5.0.0
wcwidth==0.2.5
